Установка расширения только в дев-моде, перемещение функции установки расширений в utils
