+++ src/main/index.ts
@@ -1,8 +1,7 @@
 import { is } from '@electron-toolkit/utils';
-import { app, shell, BrowserWindow, ipcMain, globalShortcut, session } from 'electron';
+import { app, shell, BrowserWindow, ipcMain, globalShortcut } from 'electron';
 import settings from 'electron-settings';
 import { lookpath } from 'lookpath';
-import { unzip } from 'unzip-crx-3';
 
 import { existsSync } from 'fs';
 import path, { join } from 'path';
@@ -19,48 +18,9 @@ import {
   settingsChangeSend,
 } from './settings';
 import { getAllTemplates, getTemplate } from './templates';
-import { basePath } from './utils';
+import { basePath, installDevToolsExtension } from './utils';
 
 import icon from '../../resources/icon.png?asset';
-
-const unzip: any = require('unzip-crx-3'); // eslint-disable-line
-
-/**
- * Устанавливает расширение для DevTools, если оно не установлено.
- * @param {string} extensionName - Название расширения (например, 'react-dev-tools')
- */
-async function installDevToolsExtension(extensionName: string) {
-  const extensionsDir = path.resolve(__dirname, '../../extensions');
-  const crxPath = path.join(extensionsDir, `${extensionName}.crx`);
-  const unpackedPath = path.join(extensionsDir, extensionName);
-
-  // Проверяем, распаковано ли расширение
-  if (!existsSync(unpackedPath)) {
-    try {
-      await unzip(crxPath, unpackedPath);
-      console.log(`Extension ${extensionName} распаковано.`);
-    } catch (err) {
-      console.error(`Ошибка распаковки ${extensionName}:`, err);
-      return;
-    }
-  }
-
-  // Проверяем, установлено ли расширение
-  const loadedExtensions = session.defaultSession.getAllExtensions?.() || [];
-  const alreadyLoaded = loadedExtensions.some((ext: any) => ext.path === unpackedPath);
-  if (!alreadyLoaded) {
-    try {
-      const ext = await session.defaultSession.loadExtension(unpackedPath, {
-        allowFileAccess: true,
-      });
-      console.log(`Loaded extension: ${ext.name}`);
-    } catch (err) {
-      console.error(`Ошибка установки расширения ${extensionName}:`, err);
-    }
-  } else {
-    console.log(`Extension ${extensionName} уже установлено.`);
-  }
-}
 
 /**
  * Создание главного окна редактора.
@@ -179,7 +139,9 @@ startModules();
 
 // Выполняется после инициализации Electron
 app.whenReady().then(() => {
-  installDevToolsExtension('react-dev-tools');
+  if (!app.isPackaged) {
+    installDevToolsExtension('react-dev-tools');
+  }
   ipcMain.handle('appVersion', app.getVersion);
 
   const mainWindow = createWindow();
