+++ src/main/utils.ts
@@ -1,3 +1,6 @@
+import { session } from 'electron';
+import { unzip } from 'unzip-crx-3';
+
 import * as fs from 'fs';
 import path from 'path';
 
@@ -41,3 +44,42 @@ export const changePermissions = (dir: string, mode: string | number) => {
     }
   });
 };
+
+const unzip: any = require('unzip-crx-3'); // eslint-disable-line
+
+/**
+ * Устанавливает расширение для DevTools, если оно не установлено.
+ * @param {string} extensionName - Название расширения (например, 'react-dev-tools')
+ */
+export async function installDevToolsExtension(extensionName: string) {
+  const extensionsDir = path.resolve(__dirname, '../../extensions');
+  const crxPath = path.join(extensionsDir, `${extensionName}.crx`);
+  const unpackedPath = path.join(extensionsDir, extensionName);
+
+  // Проверяем, распаковано ли расширение
+  if (!fs.existsSync(unpackedPath)) {
+    try {
+      await unzip(crxPath, unpackedPath);
+      console.log(`Extension ${extensionName} распаковано.`);
+    } catch (err) {
+      console.error(`Ошибка распаковки ${extensionName}:`, err);
+      return;
+    }
+  }
+
+  // Проверяем, установлено ли расширение
+  const loadedExtensions = session.defaultSession.getAllExtensions?.() || [];
+  const alreadyLoaded = loadedExtensions.some((ext: any) => ext.path === unpackedPath);
+  if (!alreadyLoaded) {
+    try {
+      const ext = await session.defaultSession.loadExtension(unpackedPath, {
+        allowFileAccess: true,
+      });
+      console.log(`Loaded extension: ${ext.name}`);
+    } catch (err) {
+      console.error(`Ошибка установки расширения ${extensionName}:`, err);
+    }
+  } else {
+    console.log(`Extension ${extensionName} уже установлено.`);
+  }
+}
